+++ src/views/UserDetails/index.js
@@ -1,12 +1,8 @@
 // @flow
 import React, { useState, useContext, useMemo, useCallback } from "react";
-import { Link, Redirect, useParams } from "react-router-dom";
-import SettingsIcon from "@material-ui/icons/Settings";
-import Button from "@material-ui/core/Button";
+import { Redirect, useRouteMatch } from "react-router-dom";
 import Grid from "@material-ui/core/Grid";
-import Snackbar from "@material-ui/core/Snackbar";
 import Typography from "@material-ui/core/Typography";
-import KeyboardBackspace from "@material-ui/icons/KeyboardBackspace";
 import { useIdle, useAudio } from "react-use";
 // $FlowFixMe
 import kaChing from "data/ka-ching.mp3";
@@ -15,44 +11,59 @@ import AsidePanel from "./AsidePanel";
 import ChangeCreditPanel from "./ChangeCreditPanel";
 import * as Cur from "components/Currency";
 import BarcodeScanner from "components/BarcodeScanner";
-import TopBar from "components/TopBar";
-import Main from "components/Main";
+import VertMenu from "components/VerticalMenu/UserDetailsVerticalMenu";
 import ProductsContext from "contexts/Products";
 import useErrorHandler from "contexts/Error";
+import { useSnackbar } from "contexts/Snackbar";
 import useUser from "hooks/useUser";
 import makeStyles from "@material-ui/styles/makeStyles";
 
 const ChangeCreditPanels = React.memo((props) => {
-  const { addCredit, backToList, onBarcodeNotFound } = props;
+  const { addCredit, search } = props;
   const products = useContext(ProductsContext);
+  const setSnackbarMsg = useSnackbar();
+  const [backToList, setBackToList] = useState(false);
   const categories = useMemo(() =>
     [...new Set(products.map((obj) => obj.category))]);
-  const scannerSuccess = (msg: string) => {
+  const scannerSuccess = useCallback((msg: string) => {
     const product = products.find((prod) => prod.ean.split("|").includes(msg));
     if (product != null) {
       addCredit(product)();
     } else if (msg === "back") {
-      backToList();
+      setBackToList(true);
+      setTimeout(() => setBackToList(false), 100);
     } else {
-      onBarcodeNotFound();
+      setSnackbarMsg(
+        <Typography color="error">
+          Unknown barcode scanned. Please enter the transaction manually.
+        </Typography>
+      );
     }
-  };
-  const filterProducts = (cat) => products.filter((p) => p.category === cat);
+  });
+  const filterProducts = useCallback((cat) => (
+    products.filter((p) => p.category === cat)
+  ));
+  const searchProducts = useCallback((cat) => (
+    filterProducts(cat).filter(
+      (p) => p.name.toLowerCase().includes(search.toLowerCase())
+    )
+  ));
   const renderedCategories = useMemo(() => (
     categories.map((cat) => (
       <ChangeCreditPanel // $FlowFixMe
-        products={filterProducts(cat)}
+        products={searchProducts(cat)}
         category={cat} key={cat} addCredit={addCredit} />
     ))
-  ), [products]);
+  ), [products, addCredit, search]);
   return (
     <Grid item xs={12} md={9}>
+      { backToList && <Redirect to="/" /> }
       <BarcodeScanner onSuccess={scannerSuccess} />
-      <ChangeCreditPanel products={[0.5, 1, 2, 5, 10]}
-        category="Add Credit" addCredit={addCredit} condensed />
+      { search !== "" || <ChangeCreditPanel products={[0.5, 1, 2, 5, 10]}
+        category="Add Credit" addCredit={addCredit} condensed /> }
       { renderedCategories }
-      <ChangeCreditPanel products={[-0.5, -1, -1.5, -2, -5]}
-        category="Remove Credit" addCredit={addCredit} condensed />
+      { search !== "" || <ChangeCreditPanel products={[-0.5, -1, -1.5, -2, -5]}
+        category="Remove Credit" addCredit={addCredit} condensed /> }
     </Grid>
   );
 });
@@ -66,33 +77,48 @@ const useStyles = makeStyles((theme) => ({
     [theme.breakpoints.up("md")]: {
       paddingRight: theme.spacing(2)
     }
+  },
+  hidden: {
+    display: "none"
   }
 }));
 
-const UserDetails = React.memo<{}>(() => {
-  const { userId } = useParams();
-  const classes = useStyles();
-  const isIdle = useIdle(30e3);
-  const [redirectBack, setBackToList] = useState(false);
+const LoadedUserDetails = React.memo(({ userId }) => {
   const user = useUser(userId);
-  const backToList = useCallback(() => setBackToList(true));
-  const [snackbarMsg, setSnackbarMsg] = useState(null);
-  const closeSnackbar = useCallback(() => setSnackbarMsg(null));
-  const handleError = useErrorHandler();
+  const isIdle = useIdle(30e3);
+  return (
+    <React.Fragment>
+      {isIdle && <Redirect to="/" /> }
+      <AsidePanel user={user} />
+    </React.Fragment>
+  );
+});
+
+type Props = {
+  vertMenuAnchorEl: Object,
+  handleCloseVertMenu: () => void,
+  search: string
+};
+
+const UserDetails = React.memo<Props>((props) => {
+  const { vertMenuAnchorEl, handleCloseVertMenu, search } = props;
+  const userDetailsMatch = useRouteMatch("/user/:userId");
+  const setSnackbarMsg = useSnackbar();
+  const classes = useStyles();
+  const uid = userDetailsMatch != null ? userDetailsMatch.params.userId : null;
   const [audio, _audioState, audioControls] = useAudio({ src: `/${kaChing}` });
   const kaching = () => {
     // reset back to 0 for when two transactions are too close after each other
     audioControls.seek(0);
     audioControls.play();
   };
-  const handleBarcodeNotFound = useCallback(() => {
-    setSnackbarMsg(
-      <Typography color="error">
-        Unknown barcode scanned. Please enter the transaction manually.
-      </Typography>
-    );
-  });
+  const handleError = useErrorHandler();
   const addCredit = useCallback((ap: Product | number) => () => {
+    if (uid == null) {
+      handleError("Something went wrong. UserId is null.");
+      return;
+    }
+    const userId: number = uid;
     if (typeof ap === "number") {
       const am: number = ap;
       API.addCredit(userId, am)
@@ -113,37 +139,20 @@ const UserDetails = React.memo<{}>(() => {
             + `for ${Cur.formatString(p.price)}`);
         }).catch(handleError);
     }
-  }, []);
+  }, [uid]);
+  const gridClass = uid == null ? classes.hidden : "";
   return (
-    <React.Fragment>
-      { (redirectBack || isIdle) && <Redirect to="/" /> }
-      <TopBar leftNode={
-        <Button component={Link} to="/">
-          <KeyboardBackspace /> Back
-        </Button>
-      } title={`User: ${user.name}`}
-      fabProps={{ component: Link, to: `/settings/${userId}` }}
-      fabIcon={<SettingsIcon />} />
-      <Main>
-        <Grid container justify="center">
-          <Grid item xs={12} md={3} className={classes.aside}>
-            <AsidePanel user={user} />
-          </Grid>
-          <ChangeCreditPanels addCredit={addCredit}
-            backToList={backToList}
-            onBarcodeNotFound={handleBarcodeNotFound} />
-        </Grid>
-      </Main>
+    <Grid container justify="center" className={gridClass}>
+      <Grid item xs={12} md={3} className={classes.aside}>
+        { uid == null ? null : <React.Fragment>
+          <LoadedUserDetails userId={uid} />
+          <VertMenu anchorEl={vertMenuAnchorEl}
+            onClose={handleCloseVertMenu} uid={uid} />
+        </React.Fragment>}
+      </Grid>
+      <ChangeCreditPanels addCredit={addCredit} search={search} />
       {audio}
-      <Snackbar
-        open={snackbarMsg != null}
-        onClose={closeSnackbar}
-        ContentProps={{
-          "aria-describedby": "message-id"
-        }}
-        message={<span id="message-id">{snackbarMsg}</span>}
-      />
-    </React.Fragment>
+    </Grid>
   );
 });